Remove no_std from ksupervisord
no_std is still not widely supported at the time of the commit.
There is also no clear way of checking for alloc errors in no_std
crates while it is instead possible to use experimental allocation
API in the standard library. There is no way to enforce it but
it might be better than using a no_std executable with its drawbacks.
In addition, linking to libc wasn't working correctly.